third_party/blink/web_tests/virtual/css-sizing-keyword-animation/external/wpt/css/css-sizing/animation/height-composition-expected.txt

This is a testharness.js-based test.
[FAIL] Compositing: property <height> underlying [100px] from add [100px] to add [auto] at (-0.3) should be [200px]
  assert_equals: expected "200px " but got "230px "
[FAIL] Compositing: property <height> underlying [100px] from add [100px] to add [auto] at (0.5) should be [auto]
  assert_equals: expected "0px " but got "150px "
[FAIL] Compositing: property <height> underlying [100px] from add [100px] to add [auto] at (1) should be [auto]
  assert_equals: expected "0px " but got "100px "
[FAIL] Compositing: property <height> underlying [100px] from add [100px] to add [auto] at (1.5) should be [auto]
  assert_equals: expected "0px " but got "50px "
Harness: the test ran to completion.
